Return currency from resource type items
- Change namespace for transformer
- Update schemas to return currency object
- Update resource type item schemas
- Update resource type item models and transformers
