+++ app/Entity/Item/SimpleExpense.php
@@ -88,7 +88,7 @@ class SimpleExpense extends Item
 
     public function resourceTypeTransformer(array $data_to_transform): Transformer
     {
-        return new \App\Models\Transformers\ResourceTypeItemType\SimpleExpense($data_to_transform);
+        return new \App\Models\Transformers\ResourceTypeItem\SimpleExpense($data_to_transform);
     }
 
     public function summaryResourceTypeModel(): Model
