+++ app/Models/Transformers/ResourceTypeItem/SimpleExpense.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Models\Transformers\ResourceTypeItemType;
+namespace App\Models\Transformers\ResourceTypeItem;
 
 use App\Models\Transformers\Transformer;
 
@@ -12,7 +12,7 @@ use App\Models\Transformers\Transformer;
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class AllocatedExpense extends Transformer
+class SimpleExpense extends Transformer
 {
     public function format(array $to_transform): void
     {
@@ -20,10 +20,12 @@ class AllocatedExpense extends Transformer
             'id' => $this->hash->item()->encode($to_transform['item_id']),
             'name' => $to_transform['item_name'],
             'description' => $to_transform['item_description'],
+            'currency' => [
+                'id' => $this->hash->currency()->encode($to_transform['item_currency_id']),
+                'code' => $to_transform['item_currency_code'],
+                'name' => $to_transform['item_currency_name'],
+            ],
             'total' => number_format((float) $to_transform['item_total'], 2, '.', ''),
-            'percentage' => (int) $to_transform['item_percentage'],
-            'actualised_total' => number_format((float) $to_transform['item_actualised_total'], 2, '.', ''),
-            'effective_date' => $to_transform['item_effective_date'],
             'created' => $to_transform['item_created_at'],
             'updated' => $to_transform['item_updated_at'],
             'resource' => [
