+++ public/api/schema/resource-type-item-simple-expense.json
@@ -10,6 +10,22 @@
         "description": {
             "type": "null"
         },
+        "currency": {
+            "type": "object",
+            "properties": {
+                "id": {
+                    "type": "string"
+                },
+                "code": {
+                    "type": "string",
+                    "maxLength": 3,
+                    "minLength": 3
+                },
+                "name": {
+                    "type": "string"
+                }
+            }
+        },
         "total": {
             "type": "string"
         },
@@ -38,6 +54,7 @@
         "id",
         "name",
         "description",
+        "currency",
         "total",
         "created",
         "updated",
